Simplification and optimisation of jobs
It comes from the use of this module that it can
be simplified to only get 1 argument. It reduces
either the complexity of the code and the time spent
in job management.

It optimise how normal jobs are called (not intermediate
function).

Also factorisation and simplification of synchronous
enter/call.

Change-Id: Id60dd2363fee2b007b4557430659faf733d2179d
Signed-off-by: José Bollo <jose.bollo@iot.bzh>