debug: Create a file indicating the waiting point
When AFB_DEBUG_WAIT is used, the file /tmp/afb-debug-<PID>
is created and contains the name of the awaiting point
(without tailing new line)

Signals other than SIGINT are ignored when waiting
for SIGINT. This improves the cleaning up. Example:
the signal SIGTERM is delayed until after removal of
the debugging indication file.

Change-Id: Iea163f87e160f9a99bc10f7365920893752bd308
Signed-off-by: José Bollo <jose.bollo@iot.bzh>