Fix const issue in afb-auth.h
The fields type and loa were declared as const.
This worked well for C compilers that are allowing
to initialize constant fixed structures.

But when compiling with C++, it is no more easy
to create initialized structure explicitely.
In C++, the solution is to use C++14 constexpr functions.
But for using these function, const must be removed.

Bug-AGL: SPEC-741

Change-Id: Ib332949b63892d9829db31d3c1a64d7d388db744
Signed-off-by: José Bollo <jose.bollo@iot.bzh>