Fix const issue in afb-auth.h
The fields type and loa were declared as const.
This worked well for C compilers that are allowing
to initialize constant fixed structures.

But when compiling with C++, it is no more easy
to create initialized structure explicitely.
In C++, the solution is to use C++14 constexpr functions.
But for using these function, const must be removed.

Bug-AGL: SPEC-741

Change-Id: I8096db6e4eba020f96fbf25789b2990d2a8cbabf
Signed-off-by: José Bollo <jose.bollo@iot.bzh>