afb-proto-ws: Split afb-stub-ws in two
The file afb-stub-ws is split in two parts:
 - afb-stub-ws: implements the interface between
   xreq and protocol.
 - afb-proto-ws: implements the pbinary protocol
   over wer sockets

The intent is to have a clean separation between
a kind of generic stub and the protocol implmentation.

This will allow soon to provide a library to access
directly in C to the websocket of an API exposed
through the otpion --ws-server.

At the moment, the protocol implementation is on
websockets but in the futur it will be cleanly
rewritten to use any binary transport.

This commit also fixe a double free in subcalls.

Change-Id: I831787533db9fe073d060dd8ee9401cbab2894e1
Signed-off-by: José Bollo <jose.bollo@iot.bzh>