APIv3: Allow to write application binding
This enable a binding version 3 to declare
no API. This is used to start a job that will
will run after initialisation.

The tutorial tuto-app1 shows how. Run it with:

  afb-daemon --binding tuto-app1.so

A further option could be add to close stdin
even if running in foreground, as it was the
case before.

Change-Id: I2b384d125accb4642eed8e004642ba959326878f
Signed-off-by: José Bollo <jose.bollo@iot.bzh>