afb-api-so-v3: Keep root API in all cases
For bindings v3, it is possible to have no
explicit default root API.

In that case, the binder creates a fake API
that references the binding itself. That api
isn't callable be it allows to emit messages,
to catch events and to call services (other
apis).

It must be kept to ensure that there is no
further error in the hypothetical case where
it is used.

Bug-AGL: SPEC-1812

Change-Id: I1a6c1e1390411a45521b9a6c1c386dc7807191df
Signed-off-by: José Bollo <jose.bollo@iot.bzh>